+++ Sources/Expression.swift
@@ -2,7 +2,7 @@
 //  Expression.swift
 //  Expression
 //
-//  Version 0.13.6
+//  Version 0.13.7
 //
 //  Created by Nick Lockwood on 15/09/2016.
 //  Copyright © 2016 Nick Lockwood. All rights reserved.